requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

aiofiles==23.1.0; python_version >= '3.7' and python_full_version < '4.0.0'

aiohttp==3.8.4; python_version >= '3.6'

aiosignal==1.3.1; python_version >= '3.7'

altair==5.0.1; python_version >= '3.7'

anyio==3.7.0; python_version >= '3.7'

async-timeout==4.0.2; python_version >= '3.6'

asyncer==0.0.2; python_full_version >= '3.6.2' and python_full_version < '4.0.0'

attrs==23.1.0; python_version >= '3.7'

certifi==2023.5.7; python_version >= '3.6'

charset-normalizer==3.1.0; python_version >= '3.7'

click==8.1.3; python_version >= '3.7'

colorama==0.4.6; platform_system == 'Windows'

coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

contourpy==1.1.0; python_version >= '3.8'

cycler==0.11.0; python_version >= '3.6'

fastapi==0.97.0; python_version >= '3.7'

ffmpy==0.3.0

filelock==3.12.2; python_version >= '3.7'

filetype==1.2.0

flatbuffers==23.5.26

fonttools==4.40.0; python_version >= '3.8'

frozenlist==1.3.3; python_version >= '3.7'

fsspec==2023.6.0; python_version >= '3.8'

gradio-client==0.2.7; python_version >= '3.8'

gradio==3.35.2; python_version >= '3.8'

h11==0.14.0; python_version >= '3.7'

httpcore==0.17.2; python_version >= '3.7'

httpx==0.24.1; python_version >= '3.7'

huggingface-hub==0.15.1; python_version >= '3.7'

humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

idna==3.4; python_version >= '3.5'

imagehash==4.3.1

imageio==2.31.1; python_version >= '3.7'

jinja2==3.1.2; python_version >= '3.7'

jsonschema==4.17.3; python_version >= '3.7'

kiwisolver==1.4.4; python_version >= '3.7'

lazy-loader==0.2; python_version >= '3.7'

linkify-it-py==2.0.2

llvmlite==0.40.1rc1; python_version >= '3.8'

markdown-it-py[linkify]==2.2.0; python_version >= '3.7'

markupsafe==2.1.3; python_version >= '3.7'

matplotlib==3.7.1; python_version >= '3.8'

mdit-py-plugins==0.3.3; python_version >= '3.7'

mdurl==0.1.2; python_version >= '3.7'

mpmath==1.3.0

multidict==6.0.4; python_version >= '3.7'

networkx==3.1; python_version >= '3.8'

numba==0.57.0; python_version >= '3.8'

numpy==1.24.3; python_version >= '3.8'

onnxruntime==1.15.0

opencv-python-headless==4.7.0.72; python_version >= '3.6'

orjson==3.9.1; python_version >= '3.7'

packaging==23.1; python_version >= '3.7'

pandas==2.0.2; python_version >= '3.8'

pillow==9.5.0

platformdirs==3.5.3; python_version >= '3.7'

pooch==1.7.0; python_version >= '3.7'

protobuf==4.23.3; python_version >= '3.7'

pydantic==1.10.9; python_version >= '3.7'

pydub==0.25.1

pygments==2.15.1; python_version >= '3.7'

pymatting==1.1.8; python_version >= '3'

pyparsing==3.0.9; python_full_version >= '3.6.8'

pyreadline3==3.4.1; python_version >= '3.8' and sys_platform == 'win32'

pyrsistent==0.19.3; python_version >= '3.7'

python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

python-multipart==0.0.6; python_version >= '3.7'

pytz==2023.3

pywavelets==1.4.1; python_version >= '3.8'

pyyaml==6.0; python_version >= '3.6'

rembg==2.0.43

requests==2.31.0; python_version >= '3.7'

scikit-image==0.21.0; python_version >= '3.8'

scipy==1.10.1; python_version < '3.12' and python_version >= '3.8'

semantic-version==2.10.0; python_version >= '2.7'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sniffio==1.3.0; python_version >= '3.7'

starlette==0.27.0; python_version >= '3.7'

sympy==1.12; python_version >= '3.8'

tifffile==2023.4.12; python_version >= '3.8'

toolz==0.12.0; python_version >= '3.5'

tqdm==4.65.0; python_version >= '3.7'

typing-extensions==4.6.3; python_version >= '3.7'

tzdata==2023.3; python_version >= '2'

uc-micro-py==1.0.2; python_version >= '3.7'

urllib3==2.0.3; python_version >= '3.7'

uvicorn==0.22.0; python_version >= '3.7'

watchdog==3.0.0; python_version >= '3.7'

websockets==11.0.3; python_version >= '3.7'

yarl==1.9.2; python_version >= '3.7'

